initial renames & tests for three important sandbox scripts
