+++ tests/test-data/paired-mixed.fa.se
@@ -1,0 +1,10 @@
+>850:2:1:1118:7944/1
+TTAATTTTGGAAACCCTGCAATAAAGTCACAACATTGC
+>850:2:1:1123:19958/1
+GCGATAAAAAGTCGTTGAGATAATCCGCGATTTCTCGCA
+>850:2:1:1126:16664/1
+CCATGTAGCGCCGCACACCTTTGTAGGTGTTGTAATAATCTTCG
+>850:2:1:1128:16434/1
+GCGGGGTCTTGCCTGCCACCCCTGGACGCCCACTGCATCCCCATGGACAC
+>850:2:1:1131:10632/1
+CGTTCAGTGAAACTTTTTCCATTGCTTTGCGCGCCGCCTCAGAGGCTTTTCGAATCGCCTC
